Match moonrock getObjGroupStatus return as u8
Cast the vtable getObjGroupStatus result to u8 so the == 0 test emits
clrlwi;cmplwi like target (recipe #11/#3 — the slot returns a byte).
mmp_moonrock_update 97.96->98.33%, unit 99.21->99.35%.

+++ src/main/dll/MMP/dll_0182_mmpmoonrock.c
@@ -697,7 +697,7 @@ void mmp_moonrock_update(int obj)
     }
     grabbed = 0;
     if ((state->flags & 8) != 0 &&
-        (*gMapEventInterface)->getObjGroupStatus(0x12, 6) == 0)
+        (u8)(*gMapEventInterface)->getObjGroupStatus(0x12, 6) == 0)
     {
         state->flags |= 1;
     }
